Create reg token for clusters with private endpoint
For EKS clusters created/imported with private-only endpoint, a cluster
registration token is generated and a URL is added to the EKSConfig
annotations to ask eks-operator to deploy a bastion to deploy the
cluster agent. This will skip the previously-required manual step of a
user running the import command on a cluster with private-only endpoint.