Skip cloud-cred check unless it's being changed
The KEv2 validators were checking the user's access to the cloud
credential every time. This would cause problems if the user was given
access to the cluster by an administrator. Instead, the logic
validates the user's access to the old cloud-credential and the new
cloud-credential only when the user is attempting to change it.